* dev-master: (36 commits)
  Lokalise updated
  lokalize updated
  New user issue fixed
  Update README.md
  Bug fixes
  Update assets_list_view_page.dart
  countries added to bank name
  add button added
  Fix: 31527
  change error message of custodian cancel
  Fix: 31517
  update ln
  Fix: onboarding
  Fix: Onboarding
  Fix: 31430
  Fix: 31430
  fix redirect issue
  Update settings_page.dart
  add new ln key and use
  Update status_step_widget.dart
  ...

+++ lib/features/valuation/data/models/post_valuation_params.dart
@@ -10,7 +10,7 @@ class PostValuationParams extends Equatable {
   final String? originCode;
   final String? wealthType;
   final DateTime? valuatedAt;
-  final int? quantity;
+  final double? quantity;
   final double? pricePerUnit;
   final double? ownershipPercentage;
   final String? note;
@@ -51,7 +51,7 @@ class PostValuationParams extends Equatable {
               ? DateTime.parse(json["valuatedAt"].toString())
               : DateTime.now(),
           quantity: json["quantity"] != null
-              ? int.tryParse(json["quantity"])
+              ? double.tryParse(json["quantity"])
               : json["quantity"],
           pricePerUnit: json["pricePerUnit"] != null
               ? double.tryParse(
